Fix word_count type
- The types of word_count and authors_note_word_count must be the same to allows
operations such as Sum to be used on both with Django's ORM queries

+++ stats/models.py
@@ -59,7 +59,7 @@ class Chapter(models.Model):
     post_date = models.DateTimeField()
     last_update = models.DateTimeField()
     download_date = models.DateTimeField()
-    word_count = models.PositiveIntegerField(default=0)
+    word_count = models.PositiveBigIntegerField(default=0)
     authors_note_word_count = models.PositiveBigIntegerField(default=0)
     book = models.ForeignKey(Book, on_delete=models.CASCADE)
